etl/transform/text-to-image/requirements.txt
Pillow==9.3.0
emoji
pilmoji